move utils to benchmark.py
